backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
httpx
python-multipart
pydantic
pydantic-settings
python-dotenv
asyncpg